[#20250122] Give vite_assets its own home
